deprecation(semver): rename `rangeFormat()` to `formatRange()` (#4090)
