+++ semver/range_format.ts
@@ -1,14 +1,16 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-import type { SemVerRange } from "./types.ts";
-import { comparatorFormat } from "./comparator_format.ts";
+
+import { formatRange } from "./format_range.ts";
+import { SemVerRange } from "./types.ts";
 
 /**
  * Formats the range into a string
  * @example >=0.0.0 || <1.0.0
  * @param range The range to format
  * @returns A string representation of the range
+ *
+ * @deprecated (will be removed after 0.213.0) Use {@linkcode formatRange} instead.
  */
 export function rangeFormat(range: SemVerRange): string {
-  return range.ranges.map((c) => c.map((c) => comparatorFormat(c)).join(" "))
-    .join("||");
+  return formatRange(range);
 }
