+++ semver/range_test.ts
@@ -1,7 +1,6 @@
 // Copyright Isaac Z. Schlueter and Contributors. All rights reserved. ISC license.
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-import { assert, assertEquals } from "../assert/mod.ts";
-import { rangeFormat } from "./range_format.ts";
+import { assert } from "../assert/mod.ts";
 import { parse } from "./parse.ts";
 import { parseRange } from "./parse_range.ts";
 import { testRange } from "./test_range.ts";
@@ -216,72 +215,4 @@ Deno.test("negativeUnlockedPrereleaseRange", function () {
     const found = testRange(s, range);
     assert(!found, `${v} satisfied by ${r} unexpectedly`);
   }
-});
-
-Deno.test({
-  name: "validRange",
-  fn: async (t) => {
-    const versions: [string, string][] = [
-      ["1.0.0 - 2.0.0", ">=1.0.0 <=2.0.0"],
-      ["1.0.0", "1.0.0"],
-      [">=*", "*"],
-      ["", "*"],
-      ["*", "*"],
-      [">=1.0.0", ">=1.0.0"],
-      [">1.0.0", ">1.0.0"],
-      ["<=2.0.0", "<=2.0.0"],
-      ["1", ">=1.0.0 <2.0.0"],
-      ["<=2.0.0", "<=2.0.0"],
-      ["<=2.0.0", "<=2.0.0"],
-      ["<2.0.0", "<2.0.0"],
-      ["<2.0.0", "<2.0.0"],
-      [">=0.1.97", ">=0.1.97"],
-      [">=0.1.97", ">=0.1.97"],
-      ["0.1.20 || 1.2.4", "0.1.20||1.2.4"],
-      [">=0.2.3 || <0.0.1", ">=0.2.3||<0.0.1"],
-      [">=0.2.3 || <0.0.1", ">=0.2.3||<0.0.1"],
-      [">=0.2.3 || <0.0.1", ">=0.2.3||<0.0.1"],
-      ["||", "*||*"],
-      ["2.x.x", ">=2.0.0 <3.0.0"],
-      ["1.2.x", ">=1.2.0 <1.3.0"],
-      ["1.2.x || 2.x", ">=1.2.0 <1.3.0||>=2.0.0 <3.0.0"],
-      ["1.2.x || 2.x", ">=1.2.0 <1.3.0||>=2.0.0 <3.0.0"],
-      ["x", "*"],
-      ["2.*.*", ">=2.0.0 <3.0.0"],
-      ["1.2.*", ">=1.2.0 <1.3.0"],
-      ["1.2.* || 2.*", ">=1.2.0 <1.3.0||>=2.0.0 <3.0.0"],
-      ["2", ">=2.0.0 <3.0.0"],
-      ["2.3", ">=2.3.0 <2.4.0"],
-      ["~2.4", ">=2.4.0 <2.5.0"],
-      ["~2.4", ">=2.4.0 <2.5.0"],
-      ["~>3.2.1", ">=3.2.1 <3.3.0"],
-      ["~1", ">=1.0.0 <2.0.0"],
-      ["~>1", ">=1.0.0 <2.0.0"],
-      ["~1.0", ">=1.0.0 <1.1.0"],
-      ["^0", ">=0.0.0 <1.0.0"],
-      ["^0.1", ">=0.1.0 <0.2.0"],
-      ["^1.0", ">=1.0.0 <2.0.0"],
-      ["^1.2", ">=1.2.0 <2.0.0"],
-      ["^0.0.1", ">=0.0.1 <0.0.2"],
-      ["^0.0.1-beta", ">=0.0.1-beta <0.0.2"],
-      ["^0.1.2", ">=0.1.2 <0.2.0"],
-      ["^1.2.3", ">=1.2.3 <2.0.0"],
-      ["^1.2.3-beta.4", ">=1.2.3-beta.4 <2.0.0"],
-      ["<1", "<1.0.0"],
-      [">=1", ">=1.0.0"],
-      ["<1.2", "<1.2.0"],
-      ["1", ">=1.0.0 <2.0.0"],
-    ];
-
-    for (const [r, expected] of versions) {
-      await t.step({
-        name: `${r} -> ${expected}`,
-        fn: () => {
-          const range = parseRange(r);
-          const actual = rangeFormat(range);
-          assertEquals(actual, expected);
-        },
-      });
-    }
-  },
 });
